+++ scripts/test3.py
@@ -1,0 +1,15 @@
+"""
+Script Dummy test
+"""
+import unittest
+
+class TestConfig1(unittest.TestCase):
+    """
+    Config modules test suit 1
+    """
+
+    def test1(self):
+        """
+        Script test1 docString
+        """
+        self.assertEqual(5,5)