+++ CHANGELOG.md
@@ -26,4 +26,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - add README.md
 ## [0.1.5] - 2025-03-13
 ### Added
-  - some unittest examples 
+  - some unittest examples (modules & scripts)